+++ CHANGELOG.md
@@ -1,18 +1,22 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.2](https://github.com/HamzaGhazouani/HGPlaceholders/releases/tag/0.2.2)
+
+* Hide the table view header and footer when the placeholder is shown 
+
 ## [0.2.1](https://github.com/HamzaGhazouani/HGPlaceholders/releases/tag/0.2.1)
 
 * Fix an issue with custom UICollectionViewLayout
 
 ## [0.2.0](https://github.com/HamzaGhazouani/HGPlaceholders/releases/tag/0.2.0)
 
-* Add UICollectionView compatibility 
+* Add UICollectionView compatibility
 
 ## [0.1.2](https://github.com/HamzaGhazouani/HGPlaceholders/releases/tag/0.1.2)
 
-* Add Carthage support 
-* Fix jazzy documentation issue 
+* Add Carthage support
+* Fix jazzy documentation issue
 
 ## [0.1.0](https://github.com/HamzaGhazouani/HGPlaceholders/releases/tag/0.1.0)